User Service - signUp 구현

+++ src/main/kotlin/com/b2/backoffice/user/model/User.kt
@@ -26,7 +26,7 @@ class User(
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id : Long? = null
+    val id : Int? = null
 }
 
 fun User.toResponse() : UserResponse{
